requirements.txt
# Gestionale Fibra - Production Dependencies
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Validation and serialization
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF and OCR
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Telegram Bot
python-telegram-bot==20.7

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Environment and configuration
python-dotenv==1.0.0

# Templating
jinja2==3.1.2

# Logging and utilities
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
ruff==0.1.6
